refactor: move novel route files into routes/novel/ subdirectory (#7)
Group 23 novel-related route files into a dedicated subdirectory,
reducing routes/ top-level entries from 49 to 27. File names drop
the redundant "novel" prefix (e.g. novelChapterRoutes.ts → novel/chapterRoutes.ts).
All import paths updated; API endpoints unchanged.
