+++ server/src/app.ts
@@ -25,11 +25,11 @@ import knowledgeRouter from "./routes/knowledge";
 import maintenanceRouter from "./routes/maintenance";
 import llmRouter from "./routes/llm";
 import novelRouter from "./routes/novel";
-import novelDirectorRouter from "./routes/novelDirector";
-import novelDecisionsRouter from "./routes/novelDecisions";
-import novelChapterSummaryRouter from "./routes/novelChapterSummary";
-import novelExportRouter from "./routes/novelExport";
-import novelWorkflowsRouter from "./routes/novelWorkflows";
+import novelDirectorRouter from "./routes/novel/director";
+import novelDecisionsRouter from "./routes/novel/decisions";
+import novelChapterSummaryRouter from "./routes/novel/chapterSummary";
+import novelExportRouter from "./routes/novel/export";
+import novelWorkflowsRouter from "./routes/novel/workflows";
 import ragRouter from "./routes/rag";
 import settingsAutoDirectorRouter from "./routes/settingsAutoDirector";
 import settingsRouter from "./routes/settings";
