+++ server/src/routes/novel/chapterEditorRoutes.ts
@@ -1,9 +1,9 @@
 import type { Router } from "express";
 import type { ApiResponse } from "@ai-novel/shared/types/api";
 import { z } from "zod";
-import { AppError } from "../middleware/errorHandler";
-import { validate } from "../middleware/validate";
-import type { NovelService } from "../services/novel/NovelService";
+import { AppError } from "../../middleware/errorHandler";
+import { validate } from "../../middleware/validate";
+import type { NovelService } from "../../services/novel/NovelService";
 
 interface RegisterNovelChapterEditorRoutesInput {
   router: Router;
