+++ server/src/routes/novel/chapterRoutes.ts
@@ -1,9 +1,9 @@
 import type { Router } from "express";
 import type { ApiResponse } from "@ai-novel/shared/types/api";
 import { z } from "zod";
-import { agentRuntime } from "../agents";
-import { validate } from "../middleware/validate";
-import type { NovelService } from "../services/novel/NovelService";
+import { agentRuntime } from "../../agents";
+import { validate } from "../../middleware/validate";
+import type { NovelService } from "../../services/novel/NovelService";
 
 interface RegisterNovelChapterRoutesInput {
   router: Router;
