+++ server/src/routes/novel/characterPreparationRoutes.ts
@@ -1,13 +1,13 @@
 import { Router } from "express";
 import type { ApiResponse } from "@ai-novel/shared/types/api";
 import { z } from "zod";
-import { llmProviderSchema } from "../llm/providerSchema";
-import { validate } from "../middleware/validate";
-import type { NovelService } from "../services/novel/NovelService";
+import { llmProviderSchema } from "../../llm/providerSchema";
+import { validate } from "../../middleware/validate";
+import type { NovelService } from "../../services/novel/NovelService";
 import {
   supplementalCharacterCandidateSchema,
   supplementalCharacterGenerationInputSchema,
-} from "../prompting/prompts/novel/characterPreparation.promptSchemas";
+} from "../../prompting/prompts/novel/characterPreparation.promptSchemas";
 
 const optionParamsSchema = z.object({
   id: z.string().trim().min(1),
