+++ server/src/routes/novel/export.ts
@@ -1,9 +1,9 @@
 import { Router } from "express";
 import { z } from "zod";
 import { NOVEL_EXPORT_FORMAT_VALUES, NOVEL_EXPORT_SCOPE_VALUES } from "@ai-novel/shared/types/novelExport";
-import { authMiddleware } from "../middleware/auth";
-import { validate } from "../middleware/validate";
-import { novelExportService } from "../services/novel/NovelExportService";
+import { authMiddleware } from "../../middleware/auth";
+import { validate } from "../../middleware/validate";
+import { novelExportService } from "../../services/novel/NovelExportService";
 
 const router = Router();
