+++ server/src/routes/novel/index.ts
@@ -1,29 +1,29 @@
 import { Router } from "express";
 import { z } from "zod";
 import { MAX_VOLUME_COUNT } from "@ai-novel/shared/types/volumePlanning";
-import { llmProviderSchema } from "../llm/providerSchema";
-import { authMiddleware } from "../middleware/auth";
-import { AppError } from "../middleware/errorHandler";
-import { NovelService } from "../services/novel/NovelService";
-import { NovelDraftOptimizeService } from "../services/novel/NovelDraftOptimizeService";
-import { chapterRuntimeRequestSchema } from "../services/novel/runtime/chapterRuntimeSchema";
-import { registerNovelBaseRoutes } from "./novelBaseRoutes";
-import { registerNovelChapterEditorRoutes } from "./novelChapterEditorRoutes";
-import { registerNovelChapterRoutes } from "./novelChapterRoutes";
-import { registerNovelChapterGenerationRoutes } from "./novelChapterGeneration";
-import { registerNovelCharacterDynamicsRoutes } from "./novelCharacterDynamicsRoutes";
-import { registerNovelCharacterPreparationRoutes } from "./novelCharacterPreparationRoutes";
-import { registerNovelCharacterResourceRoutes } from "./novelCharacterResourceRoutes";
-import { registerNovelCharacterSyncRoutes } from "./novelCharacterSyncRoutes";
-import { registerNovelFramingRoutes } from "./novelFramingRoutes";
-import { registerNovelPlanningRoutes } from "./novelPlanningRoutes";
-import { registerNovelProductionRoutes } from "./novelProductionRoutes";
-import { registerNovelReviewRoutes } from "./novelReviewRoutes";
-import { registerNovelSnapshotCharacterRoutes } from "./novelSnapshotCharacterRoutes";
-import { registerNovelStoryMacroRoutes } from "./novelStoryMacroRoutes";
-import { registerNovelStorylineRoutes } from "./novelStorylineRoutes";
-import { registerNovelVolumeRoutes } from "./novelVolumeRoutes";
-import { registerNovelWorldSliceRoutes } from "./novelWorldSliceRoutes";
+import { llmProviderSchema } from "../../llm/providerSchema";
+import { authMiddleware } from "../../middleware/auth";
+import { AppError } from "../../middleware/errorHandler";
+import { NovelService } from "../../services/novel/NovelService";
+import { NovelDraftOptimizeService } from "../../services/novel/NovelDraftOptimizeService";
+import { chapterRuntimeRequestSchema } from "../../services/novel/runtime/chapterRuntimeSchema";
+import { registerNovelBaseRoutes } from "./baseRoutes";
+import { registerNovelChapterEditorRoutes } from "./chapterEditorRoutes";
+import { registerNovelChapterRoutes } from "./chapterRoutes";
+import { registerNovelChapterGenerationRoutes } from "./chapterGeneration";
+import { registerNovelCharacterDynamicsRoutes } from "./characterDynamicsRoutes";
+import { registerNovelCharacterPreparationRoutes } from "./characterPreparationRoutes";
+import { registerNovelCharacterResourceRoutes } from "./characterResourceRoutes";
+import { registerNovelCharacterSyncRoutes } from "./characterSyncRoutes";
+import { registerNovelFramingRoutes } from "./framingRoutes";
+import { registerNovelPlanningRoutes } from "./planningRoutes";
+import { registerNovelProductionRoutes } from "./productionRoutes";
+import { registerNovelReviewRoutes } from "./reviewRoutes";
+import { registerNovelSnapshotCharacterRoutes } from "./snapshotCharacterRoutes";
+import { registerNovelStoryMacroRoutes } from "./storyMacroRoutes";
+import { registerNovelStorylineRoutes } from "./storylineRoutes";
+import { registerNovelVolumeRoutes } from "./volumeRoutes";
+import { registerNovelWorldSliceRoutes } from "./worldSliceRoutes";
 
 const router = Router();
 const novelService = new NovelService();
