+++ server/src/routes/novel/planningRoutes.ts
@@ -1,8 +1,8 @@
 import type { Router } from "express";
 import type { ApiResponse } from "@ai-novel/shared/types/api";
 import { z } from "zod";
-import { validate } from "../middleware/validate";
-import type { NovelService } from "../services/novel/NovelService";
+import { validate } from "../../middleware/validate";
+import type { NovelService } from "../../services/novel/NovelService";
 
 interface RegisterNovelPlanningRoutesInput {
   router: Router;
