+++ server/src/routes/novel/volumeRoutes.ts
@@ -1,8 +1,8 @@
 import type { Router } from "express";
 import type { ApiResponse } from "@ai-novel/shared/types/api";
 import { z } from "zod";
-import { validate } from "../middleware/validate";
-import type { NovelService } from "../services/novel/NovelService";
+import { validate } from "../../middleware/validate";
+import type { NovelService } from "../../services/novel/NovelService";
 import type { VolumePlanDocument } from "@ai-novel/shared/types/novel";
 
 interface RegisterNovelVolumeRoutesInput {
