+++ server/src/routes/novel/workflows.ts
@@ -1,11 +1,11 @@
 import { Router } from "express";
 import type { ApiResponse } from "@ai-novel/shared/types/api";
 import { z } from "zod";
-import { authMiddleware } from "../middleware/auth";
-import { validate } from "../middleware/validate";
-import { NovelDirectorService } from "../services/novel/director/NovelDirectorService";
-import { NovelWorkflowService } from "../services/novel/workflow/NovelWorkflowService";
-import { NovelWorkflowTaskAdapter } from "../services/task/adapters/NovelWorkflowTaskAdapter";
+import { authMiddleware } from "../../middleware/auth";
+import { validate } from "../../middleware/validate";
+import { NovelDirectorService } from "../../services/novel/director/NovelDirectorService";
+import { NovelWorkflowService } from "../../services/novel/workflow/NovelWorkflowService";
+import { NovelWorkflowTaskAdapter } from "../../services/task/adapters/NovelWorkflowTaskAdapter";
 
 const router = Router();
 const workflowService = new NovelWorkflowService();
